requirements.training.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0
requests>=2.31.0
ultralytics>=8.1.0
segmentation-models-pytorch>=0.3.3
azure-storage-blob>=12.23.1

# Training tooling
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
kagglehub>=0.2.0
kaggle>=1.5.0
datasets>=2.14.0
pandas>=2.0.0
seaborn>=0.12.0
roboflow>=1.1.0
albumentations>=1.4.0